requirements.txt
Flask
openai
render
gunicorn
python-dotenv
flask-cors

